Refactored EngineRenderer and all Renderer classes to use smart pointer
- Refactored EngineContext and EngineRenderer to be non-static.
- Set all Vulkan objects to VK_NULL_HANDLE when destroyed.
- Modified EngineContext to be a singleton.

+++ Engine/include/pipeline/compute_pipeline.h
@@ -5,8 +5,8 @@ namespace core {
 
 	class ComputePipeline : public Pipeline {
 	public:
-		ComputePipeline(VkDevice device, const std::string& shadername, const uint32_t& pushConstantSize = 0);
-		ComputePipeline(VkDevice device, const std::string& shadername, const std::vector<VkDescriptorSetLayout>& descSetLayouts, const uint32_t& pushConstantSize = 0);
+		ComputePipeline(VkDevice device, const std::string& shadername, const size_t& pushConstantSize = 0);
+		ComputePipeline(VkDevice device, const std::string& shadername, const std::vector<VkDescriptorSetLayout>& descSetLayouts, const size_t& pushConstantSize = 0);
 		~ComputePipeline() = default;
 
 	private:
