+++ Engine/source/pipeline/compute_pipeline.cpp
@@ -3,11 +3,11 @@
 
 namespace core {
 
-	ComputePipeline::ComputePipeline(VkDevice device, const std::string& shadername, const uint32_t& pushConstantSize) :
+	ComputePipeline::ComputePipeline(VkDevice device, const std::string& shadername, const size_t& pushConstantSize) :
 		ComputePipeline(device, shadername, std::vector<VkDescriptorSetLayout>(), pushConstantSize) {}
 	
-	ComputePipeline::ComputePipeline(VkDevice device, const std::string& shadername, const std::vector<VkDescriptorSetLayout>& descSetLayouts, const uint32_t& pushConstantSize) :
-		Pipeline(device, PipelineType::PIPELINE_TYPE_COMPUTE), m_shadername(shadername), m_pushConstantSize(pushConstantSize) {
+	ComputePipeline::ComputePipeline(VkDevice device, const std::string& shadername, const std::vector<VkDescriptorSetLayout>& descSetLayouts, const size_t& pushConstantSize) :
+		Pipeline(device, PipelineType::PIPELINE_TYPE_COMPUTE), m_shadername(shadername), m_pushConstantSize(static_cast<uint32_t>(pushConstantSize)) {
 
 		CreatePipelineLayout(descSetLayouts);
 		CreatePipeline();
